// frontend/src/api/api-client/model/noteOut.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * NoteApp API
 * A simple note app with folders and user support.
 * OpenAPI spec version: 1.0.0
 */
import type { NoteOutContent } from "./noteOutContent";
import type { NoteOutFolderId } from "./noteOutFolderId";

export interface NoteOut {
  /** @maxLength 50 */
  title: string;
  content?: NoteOutContent;
  folder_id?: NoteOutFolderId;
  is_starred?: boolean;
  id: string;
  user_id: string;
  created_at: string;
  updated_at: string;
}